api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
pydantic[email]==2.9.2
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.1
